requirements.txt
pandas==2.0.0
transformers==4.28.1
langchain==0.0.147
python-dotenv==1.0.0
loguru==0.7.0
openai==0.27.4
tensorflow==2.12.0
sentence_transformers==2.2.2
diskcache==5.4.0